app/requirements.txt
flask>=3.0
openpyxl>=3.1
pywebview>=5.0
pyinstaller>=6.0
